backend/requirements.txt
asgiref==3.6.0
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
defusedxml==0.7.1
Django==4.1.4
django-admin==2.0.1
django-allauth==0.52.0
django-cors-headers==3.13.0
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-rest-swagger==2.2.0
django-six==1.0.4
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.5
ecdsa==0.18.0
httpie==3.2.1
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.0
Pillow==9.2.0
psycopg2==2.9.5
psycopg2-binary==2.9.3
pyasn1==0.4.8
pycparser==2.21
Pygments==2.13.0
PyJWT==2.4.0
PySocks==1.7.1
python-jose==3.3.0
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rich==12.5.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
screen==1.0.1
Serializer==0.2.1
simplejson==3.17.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
stripe==5.1.0
tokenize-rt==5.0.0
typing_extensions==4.4.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.12
urlquote==2.0.0
xlwt==1.3.0
